Issue 493:bad http syntax deleting cloudsigma resources

+++ providers/cloudsigma-zrh/src/test/java/org/jclouds/cloudsigma/CloudSigmaClientLiveTest.java
@@ -308,18 +308,8 @@ public class CloudSigmaClientLiveTest {
             client.destroyServer(server.getUuid());
             client.destroyDrive(drive.getUuid());
          }
-         try {
-            client.destroyStaticIP(ip.getAddress());
-         } catch (IllegalArgumentException e) {
-            // this shouldn't happen
-            assertEquals(e.getMessage(), "resource_exception:Missing resource");
-         }
-         try {
-            client.destroyStaticIP(ip2.getAddress());
-         } catch (IllegalArgumentException e) {
-            // this shouldn't happen
-            assertEquals(e.getMessage(), "resource_exception:Missing resource");
-         }
+         client.destroyStaticIP(ip.getAddress());
+         client.destroyStaticIP(ip2.getAddress());
       }
    }
 
@@ -439,16 +429,10 @@ public class CloudSigmaClientLiveTest {
 
    @AfterGroups(groups = "live")
    protected void tearDown() {
-      try {
+      if (server != null)
          client.destroyServer(server.getUuid());
-      } catch (Exception e) {
-         // no need to check null or anything as we swallow all
-      }
-      try {
+      if (server != null)
          client.destroyDrive(drive.getUuid());
-      } catch (Exception e) {
-
-      }
       if (context != null)
          context.close();
    }
